refactor: chebyshev
