+++ tests/test_chebyshev_schedule.py
@@ -2,7 +2,7 @@ import numpy as np
 import pytest
 
 from pytorch_optimizer import get_chebyshev_schedule
-from pytorch_optimizer.optimizer.chebyshev_schedule import chebyshev_perm
+from pytorch_optimizer.lr_scheduler.chebyshev import chebyshev_perm
 
 
 def test_get_chebyshev_schedule():
